+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+ - `hasfieldname` method added for checking if a fieldname exists in an `AbstractDofHandler`/`FieldHandler`. 
+
+### Internal changes
+ - `getfielddims(::FieldHandler)`, `getfieldinterpolations(::FieldHandler)`, and `getfieldnames(::FieldHandler)` removed ([#647][github-647])
 
 ## [0.3.13] - 2023-03-23
 ### Added